Add contact details specific select option
We want to add the new text to the Contact DVLA page. There are other pages
which might also use this text such as the Contact DVSA page.

https://trello.com/c/WLEWKO6W/190-publish-deploy-final-dvla-changes

We're considering other ways in which to allow free text options and whether
other formats with a start button might benefit from a customisation option.
We're not going to do that now, because we want to get our change live.  These
options will be tracked in these Trello cards.

https://trello.com/c/3VvWDCl2/284-allow-free-text-option-for-start-button-text-in-simple-smart-answers
https://trello.com/c/02segYZw/283-allow-customisation-of-start-buttons-on-other-formats