Add example service_signin YAML file
In order to create a service_signin content item, we will need content
designers to fill out a YAML file with the required content data.  We create an
example YAML file to help guide content designers with this task.

Using a "|" on the first line "body" in the create_new_account allows us to add
content that is split over multiple lines.

However, when the file is loaded, the line breaks changed to "\n". Other
Editions in Publisher that store govspeak content store linebreaks as
"\r\n".

When we push service_sign_in content items to publishing-api we should do
the same thing.