Get the public_updated_at from the content item.
If the update_type is "minor" we don't want to change the
value of "public_updated_at" because this only applies to
major updates. But this is a required value in publishing-api,
so we have to pass something.

We get the public_updated_at date from the content store.

To be able test this logic we have decided to change how the
ServiceSigninPresenter is called. Rather than passing a
file name to the Presenter, we are assuming that a hash
of the content will be passed in from the eventual rake task.