Add choose_sign_in options field to service_sign_in_presenter
We realised that by using `hash_with_indifferent_access` when reading in our
test YAML file, we were then unable to symbolize the keys for `choose_sign_in`
options in keeping with the rest of the `render_for_publishing_api` hash. We
remove this and instead call `deep_symbolize_keys` on the YAML file.